RHOAIENG-72: Update Dockerfile for odh-notebook-controller component
This commit modifies the Dockerfile for the {odh-}notebook-controller component in the Kubeflow project. The changes include updating the base image, specifying the Golang version to 1.20, and configuring a non-root user for execution. More robust and secure build processes are now used including different steps for package installation, code copying, and building the notebook-controller.

The Dockerfile closely corresponds to what is used in OpenShift AI build. This way we will have closer match between what we release upstream and downstream.