[RHOAIENG-11895] fix(odh-nbc): put a newline between certs in case input is missing a newline (#411)
In case that the source of the certificates doesn't contain newline,
this will assure that at least one new line is between the concatenated
certificates. In case that the new line in the source is present
already, then there will be two new lines now, but it shouldn't be a
problem as it shouldn't break anything for us. This fixes [1].

* [1] https://issues.redhat.com/browse/RHOAIENG-11895