[RHOAIENG-15141] Fix failing E2E tests (#473)
There were two E2E tests in failure:
            --- FAIL: TestE2ENotebookController/update/thoth-minimal-oauth-notebook/Notebook_Statefulset_Validation_After_Update (60.28s)
            --- FAIL: TestE2ENotebookController/update/thoth-minimal-oauth-notebook/Verify_Notebook_Traffic_After_Update (10.75s)

The reason was that after the culling test, the culling configuration is
reverted, but the Notebook CR isn't updated and the annotation that was
added by the notebook controller: `kubeflow-resource-stopped` stays in
that CR and as such, the workbench instance isn't started back. This
change deletes this annotation at the end of the culler test and the
workbench is up and running and ready for the followup tests.

Another issue there was that the update test updated the workbench to a
non-existent image link, which resulted in a ImagePullBackOff Error in
the end. This change updates this link to some existing image.